Refactor profile bottom sheet
